+++ example/duckFactory.js
@@ -1,0 +1,7 @@
+// @flow
+
+import { DuckFactory } from '../src'
+
+const duck = new DuckFactory({ baseUrl: 'https://dog.ceo/api' })
+
+export default duck