+++ src/constants.js
@@ -1,0 +1,14 @@
+// @flow
+
+export const methods = Object.freeze({
+  DELETE: 'delete',
+  GET: 'get',
+  POST: 'post',
+  PUT: 'put'
+})
+
+export const statuses = Object.freeze({
+  BEGIN: 'BEGIN',
+  ERROR: 'ERROR',
+  SUCCESS: 'SUCCESS'
+})