+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+## [1.3.0]
+- Added the ability to define global configuration options using the `config` static method.
+- The `getState` method is now passed to action modifiers and request callback functions (`onSuccess` and `onError`).
+
 ## [1.2.0]
 - Added `onSuccess` and `onError` request options.